+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* This is the first release of the package. One can use this package for adding a custom bottom navigation bar to thier project.